requirements.txt
alembic==1.14.0
APScheduler==3.11.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.10.0.84
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.0
pyzbar==0.1.9
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.1.2
